requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
passlib[bcrypt]
python-dotenv
python-jose[cryptography]
alembic
email-validator
